pulled out tuning into separate function
